+++ src/APE.PostgreSQL.Teamwork.ViewModel/Postgres/PgDiffConstraints.cs
@@ -1,6 +1,7 @@
 ﻿// <copyright file="PgDiffConstraints.cs" company="APE Engineering GmbH">Copyright (c) APE Engineering GmbH. All rights reserved.</copyright>
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using APE.PostgreSQL.Teamwork.Model.PostgresSchema;
 using APE.PostgreSQL.Teamwork.Model.Utils;
 
@@ -26,25 +27,34 @@ namespace APE.PostgreSQL.Teamwork.ViewModel.Postgres
         /// <param name="foreignKey">Determines wheter forein keys should be processed.</param>
         public static void Create(StreamWriter writer, [NullGuard.AllowNull] PgSchema oldSchema, PgSchema newSchema, bool primaryKey, bool foreignKey, SearchPathHelper searchPathHelper)
         {
-            foreach (PgTable newTable in newSchema.Tables)
+            foreach (var newTable in newSchema.Tables)
             {
-                PgTable oldTable;
-
-                if (oldSchema == null)
-                {
-                    oldTable = null;
-                }
-                else
-                {
+                PgTable oldTable = null;
+                if (oldSchema != null)
                     oldTable = oldSchema.GetTable(newTable.Name);
-                }
+
+                var oldConstraints = new List<PgConstraint>(oldTable.Constraints);
 
                 // Add new constraints
-                foreach (PgConstraint constraint in GetNewConstraints(oldTable, newTable, primaryKey, foreignKey))
+                foreach (var newConstraint in GetNewConstraints(oldTable, newTable, primaryKey, foreignKey))
                 {
+                    if (oldTable != null)
+                    {
+                        var oldConstraint = oldConstraints.FirstOrDefault(c => c.IsRenamed(newConstraint));
+                        if (oldConstraint != null)
+                        {
+                            // constraint was renamed
+                            searchPathHelper.OutputSearchPath(writer);
+                            writer.WriteLine();
+                            writer.WriteLine(oldConstraint.RenameToSql(newConstraint.Name));
+                            oldConstraints.Remove(oldConstraint);
+                            continue;
+                        }
+                    }
+
                     searchPathHelper.OutputSearchPath(writer);
                     writer.WriteLine();
-                    writer.WriteLine(constraint.CreationSQL);
+                    writer.WriteLine(newConstraint.CreationSQL);
                 }
             }
         }
@@ -59,21 +69,14 @@ namespace APE.PostgreSQL.Teamwork.ViewModel.Postgres
         /// <param name="oldSchema">The schema of the old database.</param>
         public static void Drop(StreamWriter writer, [NullGuard.AllowNull] PgSchema oldSchema, PgSchema newSchema, bool primaryKey, SearchPathHelper searchPathHelper)
         {
-            foreach (PgTable newTable in newSchema.Tables)
+            foreach (var newTable in newSchema.Tables)
             {
-                PgTable oldTable;
-
-                if (oldSchema == null)
-                {
-                    oldTable = null;
-                }
-                else
-                {
+                PgTable oldTable = null;
+                if (oldSchema != null)
                     oldTable = oldSchema.GetTable(newTable.Name);
-                }
 
                 // Drop constraints that no more exist or are modified
-                foreach (PgConstraint constraint in GetDropConstraints(oldTable, newTable, primaryKey))
+                foreach (var constraint in GetDropConstraints(oldTable, newTable, primaryKey))
                 {
                     searchPathHelper.OutputSearchPath(writer);
                     writer.WriteLine();
@@ -160,24 +163,58 @@ namespace APE.PostgreSQL.Teamwork.ViewModel.Postgres
         }
 
         /// <summary>
+        /// Returns a list of all constraints which were renamed.
+        /// </summary>
+        /// <param name="oldTable">Original table or null.</param>
+        /// <param name="newTable">New table or null.</param>
+        /// <returns>A <see cref="Dictionary{PgConstraint, String}"/> which contains the original <see cref="PgConstraint"/> and the new <see cref="PgConstraint.Name"/>.</returns>
+        private static Dictionary<PgConstraint, string> GetRenameConstraints([NullGuard.AllowNull] PgTable oldTable, [NullGuard.AllowNull] PgTable newTable)
+        {
+            var retval = new Dictionary<PgConstraint, string>();
+
+            if (newTable != null && oldTable != null)
+            {
+                var oldConstraints = new List<PgConstraint>(oldTable.Constraints);
+                foreach (var newConstraint in newTable.Constraints)
+                {
+                    // if there are multiple constraints rename only the first
+                    var oldConstraint = oldConstraints.FirstOrDefault(c => c.IsRenamed(newConstraint));
+                    if (oldConstraint != null)
+                    {
+                        retval.Add(oldConstraint, oldConstraint.Name);
+                        oldConstraints.Remove(oldConstraint);
+                    }
+                }
+            }
+
+            return retval;
+        }
+
+        /// <summary>
         /// Returns list of constraints that should be dropped.
         /// </summary>
         /// <param name="oldTable">Original table or null.</param>
         /// <param name="newTable">New table or null.</param>
         /// <param name="primaryKey">Determines whether primary keys should be processed or any other constraints should be processed.</param>
         /// <returns>List of constraints that should be dropped.</returns>
-        private static IList<PgConstraint> GetDropConstraints([NullGuard.AllowNull] PgTable oldTable, [NullGuard.AllowNull] PgTable newTable, bool primaryKey)
+        private static List<PgConstraint> GetDropConstraints([NullGuard.AllowNull] PgTable oldTable, [NullGuard.AllowNull] PgTable newTable, bool primaryKey)
         {
             // todo db Constraints that are depending on a removed field should not be added to drop because they are already removed.
-            IList<PgConstraint> constraints = new List<PgConstraint>();
+            var constraints = new List<PgConstraint>();
+            var renamedConstraints = GetRenameConstraints(oldTable, newTable);
 
             if (newTable != null && oldTable != null)
             {
-                foreach (PgConstraint constraint in oldTable.Constraints)
+                foreach (var oldConstraint in oldTable.Constraints)
                 {
-                    if (constraint.PrimaryKeyConstraint == primaryKey && (!newTable.ContainsConstraint(constraint.Name) || !newTable.GetConstraint(constraint.Name).Equals(constraint)))
+                    // we don't have to drop this constraint because its renamed
+                    if (renamedConstraints.ContainsKey(oldConstraint))
+                        continue;
+
+                    if (oldConstraint.PrimaryKeyConstraint == primaryKey
+                        && (!newTable.ContainsConstraint(oldConstraint.Name) || !newTable.GetConstraint(oldConstraint.Name).Equals(oldConstraint)))
                     {
-                        constraints.Add(constraint);
+                        constraints.Add(oldConstraint);
                     }
                 }
             }
@@ -188,32 +225,32 @@ namespace APE.PostgreSQL.Teamwork.ViewModel.Postgres
         /// <summary>
         /// Returns list of constraints that should be added.
         /// </summary>
-        private static IList<PgConstraint> GetNewConstraints([NullGuard.AllowNull] PgTable oldTable, [NullGuard.AllowNull] PgTable newTable, bool primaryKey, bool foreignKey)
+        private static List<PgConstraint> GetNewConstraints([NullGuard.AllowNull] PgTable oldTable, [NullGuard.AllowNull] PgTable newTable, bool primaryKey, bool foreignKey)
         {
-            IList<PgConstraint> constraints = new List<PgConstraint>();
+            var constraints = new List<PgConstraint>();
+            var renamedConstraints = GetRenameConstraints(oldTable, newTable);
 
+            // no new constraints
             if (newTable == null)
-            {
                 return constraints;
-            }
 
             if (oldTable == null)
             {
-                foreach (PgConstraint constraint in newTable.Constraints)
+                // all constraints are new
+                foreach (var newConstraint in newTable.Constraints)
                 {
-                    if (constraint.PrimaryKeyConstraint == primaryKey && constraint.ForeignKeyConstraint == foreignKey)
-                    {
-                        constraints.Add(constraint);
-                    }
+                    if (newConstraint.PrimaryKeyConstraint == primaryKey && newConstraint.ForeignKeyConstraint == foreignKey)
+                        constraints.Add(newConstraint);
                 }
             }
             else
             {
-                foreach (PgConstraint constraint in newTable.Constraints)
+                foreach (var newConstraint in newTable.Constraints)
                 {
-                    if ((constraint.PrimaryKeyConstraint == primaryKey && constraint.ForeignKeyConstraint == foreignKey) && (!oldTable.ContainsConstraint(constraint.Name) || !oldTable.GetConstraint(constraint.Name).Equals(constraint)))
+                    if ((newConstraint.PrimaryKeyConstraint == primaryKey && newConstraint.ForeignKeyConstraint == foreignKey)
+                        && (!oldTable.ContainsConstraint(newConstraint.Name) || !oldTable.GetConstraint(newConstraint.Name).Equals(newConstraint)))
                     {
-                        constraints.Add(constraint);
+                        constraints.Add(newConstraint);
                     }
                 }
             }